fix data type conversion related compiler warnings
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@36899 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/support/AppleSpeller.m
@@ -115,7 +115,7 @@ SpellCheckResult AppleSpeller_check(AppleSpeller speller, const char * word, con
 					result = SPELL_CHECK_LEARNED;
 			}
 		} else {
-			int capacity = [speller->misspelled count] + 1;
+			NSUInteger capacity = [speller->misspelled count] + 1;
 			NSMutableArray * misspelled = [NSMutableArray arrayWithCapacity:capacity];
 			[misspelled addObjectsFromArray:speller->misspelled];
 			[misspelled addObject:[NSValue valueWithRange:match]];
@@ -233,7 +233,7 @@ int AppleSpeller_numMisspelledWords(AppleSpeller speller)
 
 void AppleSpeller_misspelledWord(AppleSpeller speller, int index, int * start, int * length)
 {
-	NSRange range = [[speller->misspelled objectAtIndex:index] rangeValue];
+	NSRange range = [[speller->misspelled objectAtIndex:(NSUInteger)index] rangeValue];
 	*start = range.location;
 	*length = range.length;
 }
